Improved the nested layout demo.
