+++ samples/dashboard/demos/dashboard-layout-nested/demo.html
@@ -1,1 +1,6 @@
+<script src="https://code.highcharts.com/dashboard.js"></script>
+<script src="https://code.highcharts.com/highcharts.js"></script>
+<script src="https://code.highcharts.com/modules/accessibility.js"></script>
+<script src="https://code.highcharts.com/modules/dashboard-plugin.js"></script>
+
 <div id="container" class="highcharts-dashboard"></div>
